Created RecordingManager and figured out how to disable input during playback
RecordingManager - new class that coordinates recording among various
recorded objects.
MotionRecorder - new name for class that can be added to the camera to
record position and orientation.  Put it on the cube for now as well.
To do: See if it's possible to create another class that records input
events for the attached object so that can be played back too.