Got recording to work in a round-about way
Even with the old SDK, the recording manager still was not able to find
any "GvrHead" objects in the scene to disable when playback started.
Maybe they are created in another thread or something.  What I ended up
doing whas adding some code to GvrHead to make it check
RecordingManager.State and not update if the state is Playing.  This is
not ideal, will have to investigate further.