Added a Focusable class and added SelectionMask to CustomizedGazeInputModule
Added a Focusable class that can serve as a base class for anythign
clickable.  Highlighting the object it is attached to (provided it has a
physics collider) causes all child objects to be highlighted in yellow
with an outline using a shader.

Added a SelectionMask to CustomizedGazeInputModule. This lets us
programatically set layers that are eligable for input.  For example, in
the code that Christian is working on, he could set the mask to only
allow input on the gazed object and its children by changing their layer
and then removing the default layer from the seletion mask.