Finished Pickup class
It is now possible to specify a target transform by placing an object in
the scene and linking it to the pickup class.  When the user collects a
pickup, it will display a message and first move to the camera, and then
move to the final destination.  This way the user will understand that
the table is storing their inventory.