Added Openable class and created example
Edited Globe asset in Maya to break it into individual movable parts.
Added Openable class which allows an open position to be specified which
a sub-object will animate to when opened.  When clicked again, it will
move back to the initial (closed) position.
Demo of this added to globe in ShaderTest scene.
Also updated Pickup to change parent of object to inventory position
when it is picked up.  This way if it is, say, inside a drawer and the
drawer is closed, it will not move with the drawer even though it is now
visually on the table.