Implemented Openable and Lockable
Added Openable, Lockable, and LockedByPickup classes.
LockedByPickup unlocks the object once the pickup that is placed in the
reference slot is collected.
Added examples to ShaderTest scene.  Click on the pillows to "open" them
and find the key.  Then you can open the globe, which is locked by the
key.