Finished light puzzle
Creaed a vintage lightswitch mesh using a photo from Wikipedia as a
texture.
Added behavior to globe so that clicking the countries in the right
order opens the globe.
Added behavior to the lightswitch so that it turns the lights of the
room on and off.
Added behavior to the desklamp so that clicking it turns it on and off.