Worked on a new selection glow effect
The effect can be seen in the ShaderTest2 scene.  It works by rendering
the object in a solid orange color to an off-screen image, blurring it,
and then blending it back with the original image.
Will take some additional effort to get it to work with VR eyes since it
needs to be added to both left and right eyes and the render mode cannot
be "direct."  Not sure how it will affect performance.