Made recording/playback start when the scene is truly loaded
Implemented by checking the "time since level loaded" value in the
RecordingManager update function.  This value appears to remain at zero
until the first frame is rendered, so I added a SceneFullyLoaded event
to RecordingManager that MenuManager can subscribe to to get notified
when the scene is REALLY loaded and things are moving on the screen.